+++ niimpy/analysis/__init__.py
