+++ niimpy/exploration/EDA_categorical.py
