+++ niimpy/exploration/setup_dataframe.py
