+++ niimpy/exploration/test_EDA_categorical.py
