+++ niimpy/preprocessing/filter.py
