+++ niimpy/preprocessing/subject_selection.py
