+++ niimpy/reading/read.py
